+++ pywatts/core/base.py
@@ -137,6 +137,9 @@ class Base(ABC):
         """
         return self.fit(**kwargs)
 
+    def get_min_data(self):
+        return 0
+
     def __call__(self,
                  use_inverse_transform: bool = False,
                  use_prob_transform: bool = False,
@@ -144,7 +147,7 @@ class Base(ABC):
                  condition: Optional[Callable] = None,
                  computation_mode: ComputationMode = ComputationMode.Default,
                  batch_size: Optional[pd.Timedelta] = None,
-                 train_if: Optional[Union[Callable, bool]] = None,
+                 refit_conditions: List[Union[Callable, bool]] = [],
                  lag: Optional[int] = pd.Timedelta(hours=0),
                  retrain_batch: Optional[int] = pd.Timedelta(hours=24),
                  **kwargs: Union[StepInformation, Tuple[StepInformation, ...]]
@@ -169,8 +172,9 @@ class Base(ABC):
         :type use_prob_transform: bool
         :param callbacks: Callbacks to use after results are processed.
         :type callbacks: List[BaseCallback, Callable[[Dict[str, xr.DataArray]]]]
-        :param train_if: A callable, which contains a condition that indicates if the module should be trained or not
-        :type train_if: Optional[Callable]
+        :param refit_conditions: A List of Callables of BaseConditions, which contains a condition that indicates if
+                                 the module should be trained or not
+        :type refit_conditions: List[Union[BaseCondition, Callable]]
         :param batch_size: Determines how much data from the past should be used for training
         :type batch_size: pd.Timedelta
         :param computation_mode: Determines the computation mode of the step. Could be ComputationMode.Train,
@@ -194,7 +198,9 @@ class Base(ABC):
                                          condition=condition,
                                          callbacks=callbacks,
                                          computation_mode=computation_mode, batch_size=batch_size,
-                                         train_if=train_if,
+                                         refit_conditions=refit_conditions if isinstance(refit_conditions, list) else [
+                                             refit_conditions
+                                         ],
                                          retrain_batch=retrain_batch,
                                          lag=lag
                                          )