+++ pywatts/core/pipeline.py
@@ -80,16 +80,16 @@ class Pipeline(BaseTransformer):
                 # _transform is called directly.
                 return self._transform(x, None)
             else:
-                return self._comp(x, self.current_run_setting.summary_formatter, self.batch)[0]
+                return self._comp(x, False, self.current_run_setting.summary_formatter, self.batch)
         else:
-            return self._comp(x, self.current_run_setting.summary_formatter, self.batch)[0]
+            return self._comp(x, False, self.current_run_setting.summary_formatter, self.batch)
 
     def _transform(self, x, batch=None):
         for key, (start_step, _) in self.start_steps.items():
             if not start_step.buffer:
                 start_step.buffer = {key: x[key].copy()}
             else:
-                dim = _get_time_indexes(x)[0]
+                dim = _get_time_indexes(start_step.buffer[key])[0]
                 last = start_step.buffer[key][dim].values[-1]
                 start_step.buffer[key] = xr.concat([start_step.buffer[key], x[key][x[key][dim] > last]], dim=dim)
             start_step.finished = True
@@ -98,19 +98,20 @@ class Pipeline(BaseTransformer):
         last_steps = list(filter(lambda x: x.last, self.id_to_step.values()))
         if not batch:
             return self._collect_results(last_steps)
-        return self._collect_batches(last_steps, time_index)
+        return self._collect_batches(last_steps)
 
-    def _collect_batches(self, last_steps, time_index):
+    def _collect_batches(self, last_steps):
         result = dict()
         while all(map(lambda step: step.further_elements(self.counter), last_steps)):
             print(self.counter)
             if not result:
-                result = self._collect_results(last_steps)
+                result = self._collect_results(last_steps, use_batch=not self.batch is None)
             else:
-                input_results = self._collect_results(last_steps)
+                input_results = self._collect_results(last_steps, use_batch=not self.batch is None)
                 if input_results is not None:
+                    dim = _get_time_indexes(input_results)[0]
                     for key in input_results.keys():
-                        result[key] = xr.concat([result[key], input_results[key]], dim=time_index[0])
+                        result[key] = xr.concat([result[key], input_results[key]], dim=dim)
                 else:
                     message = f"From {self.counter} until {self.counter + self.batch} no data are calculated"
                     warnings.warn(message)
@@ -119,9 +120,9 @@ class Pipeline(BaseTransformer):
             self.counter += self.batch
         return result
 
-    def _collect_results(self, inputs):
+    def _collect_results(self, inputs, use_batch=False):
         # Note the return value is None if none of the inputs provide a result for this step...
-        end = None if not self.batch else self.counter + self.batch
+        end = None if not use_batch else self.counter + self.batch
         result = dict()
         for i, step in enumerate(inputs):
             if not isinstance(step, SummaryStep):
@@ -167,7 +168,7 @@ class Pipeline(BaseTransformer):
 
         # TODO built the graph which should be drawn by starting with the last steps...
 
-    def test(self, data: Union[pd.DataFrame, xr.Dataset], summary: bool = False,
+    def test(self, data: Union[pd.DataFrame, xr.Dataset], summary: bool = True,
              summary_formatter: SummaryFormatter = SummaryMarkdown(), online_start=None):
         """
         Executes all modules in the pipeline in the correct order. This method call only transform on every module
@@ -185,7 +186,7 @@ class Pipeline(BaseTransformer):
         """
         return self._run(data, ComputationMode.Transform, summary, summary_formatter, online_start)
 
-    def train(self, data: Union[pd.DataFrame, xr.Dataset], summary: bool = False,
+    def train(self, data: Union[pd.DataFrame, xr.Dataset], summary: bool = True,
               summary_formatter: SummaryFormatter = SummaryMarkdown()):
         """
         Executes all modules in the pipeline in the correct order. This method calls fit and transform on each module
@@ -218,43 +219,41 @@ class Pipeline(BaseTransformer):
 
         if isinstance(data, pd.DataFrame):
             data = data.to_xarray()
+            data = {key: data[key] for key in data.data_vars}
+        elif isinstance(data, xr.Dataset):
+            data = {key: data[key] for key in data.data_vars}
         elif isinstance(data, dict):
             for key in data:
                 if not isinstance(data[key], xr.DataArray):
                     raise WrongParameterException(
                         "Input Dict does not contain xr.DataArray objects.",
                         "Make sure to pass Dict[str, xr.DataArray].",
-                        self.name
-                    )
-            data = xr.Dataset(data)
+                        self.name)
+        else:
+            raise WrongParameterException(
+                "Unkown data type to pass to pipeline steps.",
+                "Make sure to use pandas DataFrames, xarray Datasets, or Dict[str, xr.DataArray].",
+                self.name)
 
-        if isinstance(data, xr.Dataset):
-            if self.current_run_setting.online_start is not None:
-                # First only _transform should be called (no summary, no online) on the data before online_start.
-                # Afterwards, comp is called (_transform and summaries using online simulation)
-                index_name = _get_time_indexes(data)[0]
-                self._transform({key: data[key].sel(
-                    **{index_name: data[key][index_name] < self.current_run_setting.online_start }) for key in data.data_vars}, False)
-                for step in self.id_to_step.values():
-                    step.reset(keep_buffer=True)
-                    step.set_run_setting(self.current_run_setting.clone())
-                return self._comp({key: data[key].sel(
-                    **{index_name: data[key][index_name] >= self.current_run_setting.online_start}) for key in data.data_vars}
-                    , summary_formatter, self.batch, start=self.current_run_setting.online_start)
-            else:
-                return self._comp({key: data[key] for key in data.data_vars}, summary_formatter, self.batch)
+        if self.current_run_setting.online_start is not None:
+            # First only _transform should be called (no summary, no online) on the data before online_start.
+            # Afterwards, comp is called (_transform and summaries using online simulation)
+            index_name = _get_time_indexes(data)[0]
+            self._transform({key: data[key].sel(
+                **{index_name: data[key][index_name] < self.current_run_setting.online_start }) for key in data}, False)
+            for step in self.id_to_step.values():
+                step.reset(keep_buffer=True)
+                step.set_run_setting(self.current_run_setting.clone())
+            return self._comp({key: data[key].sel(**{index_name: data[key][index_name] >= self.current_run_setting.online_start}) for key in data},
+                              self.current_run_setting.return_summary, summary_formatter, self.batch, start=self.current_run_setting.online_start)
+        else:
+            return self._comp(data, self.current_run_setting.return_summary, summary_formatter, self.batch)
 
-        raise WrongParameterException(
-            "Unkown data type to pass to pipeline steps.",
-            "Make sure to use pandas DataFrames, xarray Datasets, or Dict[str, xr.DataArray].",
-            self.name
-        )
 
-    def _comp(self, data, summary_formatter, batch, start=None):
+    def _comp(self, data, return_summary, summary_formatter, batch, start=None):
         result = self._transform(data, batch)
         summary = self._create_summary(summary_formatter, start)
-        return result, summary
-
+        return (result, summary) if return_summary else result
 
     def add(self, *,
             module: Union[BaseStep],