+++ pywatts/core/step.py
@@ -4,11 +4,13 @@ import warnings
 from typing import Optional, Dict, Union, Callable, List
 
 import cloudpickle
+import numpy as np
 import pandas as pd
 import xarray as xr
 
 from pywatts.callbacks import BaseCallback
 from pywatts.core.base import Base, BaseEstimator
+from pywatts.core.base_condition import BaseCondition
 from pywatts.core.base_step import BaseStep
 from pywatts.core.run_setting import RunSetting
 from pywatts.core.computation_mode import ComputationMode
@@ -39,8 +41,9 @@ class Step(BaseStep):
     :type callbacks: List[Union[BaseCallback, Callable[[Dict[str, xr.DataArray]], None]]]
     :param condition: A callable which checks if the step should be executed with the current data.
     :type condition: Callable[xr.DataArray, xr.DataArray, bool]
-    :param train_if: A callable which checks if the train_if step should be executed or not.
-    :type train_if: Callable[xr.DataArray, xr.DataArray, bool]
+    :param refit_conditions: A List of Callables of BaseConditions, which contains a condition that indicates if
+                                 the module should be trained or not
+    :type refit_conditions: List[Union[BaseCondition, Callable]]
     :param lag: Needed for online learning. Determines what data can be used for retraining.
             E.g., when 24 hour forecasts are performed, a lag of 24 hours is needed, else the retraining would
             use future values as target values.
@@ -55,7 +58,7 @@ class Step(BaseStep):
                  callbacks: List[Union[BaseCallback, Callable[[Dict[str, xr.DataArray]], None]]] = [],
                  condition=None,
                  batch_size: Optional[None] = None,
-                 train_if=None,
+                 refit_conditions=[],
                  retrain_batch=pd.Timedelta(hours=24),
                  lag=pd.Timedelta(hours=24)):
         super().__init__(input_steps, targets, condition=condition,
@@ -65,13 +68,13 @@ class Step(BaseStep):
         self.retrain_batch = retrain_batch
         self.callbacks = callbacks
         self.batch_size = batch_size
-        if self.current_run_setting.computation_mode is not ComputationMode.Refit and train_if is not None:
+        if self.current_run_setting.computation_mode is not ComputationMode.Refit and len(refit_conditions) > 0:
             message = "You added a refit_condition without setting the computation_mode to refit." \
                       " The condition will be ignored."
             warnings.warn(message)
             logger.warning(message)
         self.lag = lag
-        self.train_if = train_if
+        self.refit_conditions = refit_conditions
         self.result_steps: Dict[str, ResultStep] = {}
 
     def _fit(self, inputs: Dict[str, BaseStep], target_step):
@@ -120,11 +123,11 @@ class Step(BaseStep):
                 condition = cloudpickle.load(pickle_file)
         else:
             condition = None
-        if stored_step["train_if"]:
-            with open(stored_step["train_if"], 'rb') as pickle_file:
-                train_if = cloudpickle.load(pickle_file)
-        else:
-            train_if = None
+        refit_conditions = []
+        for refit_condition in stored_step["refit_conditions"]:
+            with open(refit_condition, 'rb') as pickle_file:
+                refit_conditions.append(cloudpickle.load(pickle_file))
+
         callbacks = []
         for callback_path in stored_step["callbacks"]:
             with open(callback_path, 'rb') as pickle_file:
@@ -133,7 +136,7 @@ class Step(BaseStep):
             callbacks.append(callback)
 
         step = cls(module, inputs, targets=targets, file_manager=file_manager, condition=condition,
-                   train_if=train_if, callbacks=callbacks, batch_size=stored_step["batch_size"])
+                   refit_conditions=refit_conditions, callbacks=callbacks, batch_size=stored_step["batch_size"])
         step.default_run_setting = RunSetting.load(stored_step["default_run_setting"])
         step.current_run_setting = step.default_run_setting.clone()
         step.id = stored_step["id"]
@@ -142,15 +145,15 @@ class Step(BaseStep):
 
         return step
 
-    def _compute(self, start, end):
-        input_data = self._get_input(start, end)
-        target = self._get_target(start, end)
+    def _compute(self, start, end, minimum_data):
+        input_data = self._get_input(start, end, minimum_data)
+        target = self._get_target(start, end, minimum_data)
         if self.current_run_setting.computation_mode in [ComputationMode.Default, ComputationMode.FitTransform,
                                                          ComputationMode.Train]:
             # Fetch input_data and target data
             if self.batch_size:
-                input_batch = self._get_input(end - self.batch_size, end)
-                target_batch = self._get_target(end - self.batch_size, end)
+                input_batch = self._get_input(end - self.batch_size, end, minimum_data)
+                target_batch = self._get_target(end - self.batch_size, end, minimum_data)
                 start_time = time.time()
                 self._fit(input_batch, target_batch)
                 self.training_time.set_kv("", time.time() - start_time)
@@ -171,29 +174,42 @@ class Step(BaseStep):
             result_dict[key] = res.sel(**{_get_time_indexes(res)[0]: index[(index >= start)]})
         return result_dict
 
-    def _get_target(self, start, batch):
+    def _get_target(self, start, batch, minimum_data=(0, pd.Timedelta(0))):
+        min_data_module = self.module.get_min_data()
+        if isinstance(min_data_module, (int, np.integer)):
+            minimum_data = minimum_data[0] + min_data_module, minimum_data[1]
+        else:
+            minimum_data = minimum_data[0], minimum_data[1] + min_data_module
         return {
-            key: target.get_result(start, batch) for key, target in self.targets.items()
+            key: target.get_result(start, batch, minimum_data=minimum_data)
+            for key, target in self.targets.items()
         }
 
-    def _get_input(self, start, batch):
+    def _get_input(self, start, batch, minimum_data=(0, pd.Timedelta(0))):
+        min_data_module = self.module.get_min_data()
+        if isinstance(min_data_module, (int, np.integer)):
+            minimum_data = minimum_data[0] + min_data_module, minimum_data[1]
+        else:
+            minimum_data = minimum_data[0], minimum_data[1] + min_data_module
         return {
-            key: input_step.get_result(start, batch) for key, input_step in self.input_steps.items()
+            key: input_step.get_result(start, batch, minimum_data=minimum_data) for
+            key, input_step in self.input_steps.items()
         }
 
     def get_json(self, fm: FileManager):
         json = super().get_json(fm)
         condition_path = None
-        train_if_path = None
+        refit_conditions_paths = []
         callbacks_paths = []
         if self.condition:
             condition_path = fm.get_path(f"{self.name}_condition.pickle")
             with open(condition_path, 'wb') as outfile:
                 cloudpickle.dump(self.condition, outfile)
-        if self.train_if:
-            train_if_path = fm.get_path(f"{self.name}_train_if.pickle")
-            with open(train_if_path, 'wb') as outfile:
-                cloudpickle.dump(self.train_if, outfile)
+        for refit_condition in self.refit_conditions:
+            refit_conditions_path = fm.get_path(f"{self.name}_refit_conditions.pickle")
+            with open(refit_conditions_path, 'wb') as outfile:
+                cloudpickle.dump(refit_condition, outfile)
+            refit_conditions_paths.append(refit_conditions_path)
         for callback in self.callbacks:
             callback_path = fm.get_path(f"{self.name}_callback.pickle")
             with open(callback_path, 'wb') as outfile:
@@ -201,7 +217,7 @@ class Step(BaseStep):
             callbacks_paths.append(callback_path)
         json.update({"callbacks": callbacks_paths,
                      "condition": condition_path,
-                     "train_if": train_if_path,
+                     "refit_conditions": refit_conditions_paths,
                      "batch_size": self.batch_size})
         return json
 
@@ -213,13 +229,24 @@ class Step(BaseStep):
         """
         if self.current_run_setting.computation_mode in [ComputationMode.Refit] and isinstance(self.module,
                                                                                                BaseEstimator):
-            if self.train_if:
-                input_data = self._get_input(start, end)
-                target = self._get_target(start, end)
-                if self.train_if(input_data, target):
-                    refit_input = self._get_input(end - self.retrain_batch, end)
-                    refit_target = self._get_target(end - self.retrain_batch, end)
-                    self.module.refit(**refit_input, **refit_target)
+            for refit_condition in self.refit_conditions:
+                if isinstance(refit_condition, BaseCondition):
+                    condition_input = {key: value.step.get_result(start, end) for key, value in
+                                       refit_condition.kwargs.items()}
+                    if refit_condition.evaluate(**condition_input):
+                        self._refit(end)
+                        break
+                elif isinstance(refit_condition, Callable):
+                    input_data = self._get_input(start, end)
+                    target = self._get_target(start, end)
+                    if refit_condition(input_data, target):
+                        self._refit(end)
+                        break
+
+    def _refit(self, end):
+        refit_input = self._get_input(end - self.retrain_batch, end)
+        refit_target = self._get_target(end - self.retrain_batch, end)
+        self.module.refit(**refit_input, **refit_target)
 
     def get_result_step(self, item: str):
         if item not in self.result_steps: