+++ pywatts/modules/wrappers/sklearn_wrapper.py
@@ -11,7 +11,7 @@ from pywatts.core.exceptions.kind_of_transform_does_not_exist_exception import K
     KindOfTransform
 from pywatts.core.filemanager import FileManager
 from pywatts.utils._split_kwargs import split_kwargs
-from pywatts.utils._xarray_time_series_utils import _get_time_indexes
+from pywatts.utils._xarray_time_series_utils import _get_time_indexes, numpy_to_xarray
 from pywatts.modules.wrappers.base_wrapper import BaseWrapper
 
 
@@ -83,20 +83,17 @@ class SKLearnWrapper(BaseWrapper):
     @staticmethod
     def _sklearn_output_to_dataset(kwargs: xr.DataArray, prediction, targets: List[Tuple[str, int]]):
         reference = kwargs[list(kwargs)[0]]
-        time_index = reference.indexes[_get_time_indexes(reference)[0]]
         if len(targets) == 0:  # sklearn module is a transformer
-            coords = (
-                # first dimension is number of batches. We assume that this is the time.
-                ("time", time_index.values),
-                *[(f"dim_{j}", list(range(size))) for j, size in enumerate(prediction.shape[1:])])
-            result = xr.DataArray(prediction, coords=coords)
+            result = numpy_to_xarray(x=prediction, reference=reference)
         else:  # sklearn module is an estimator
             result = {}
             position = 0
             prediction = prediction.reshape(len(list(reference.coords.values())[0]), -1)
+            time_index = reference.indexes[_get_time_indexes(reference)[0]]
             for i, target in enumerate(targets):
                 result[target[0]] = xr.DataArray(prediction[:, position: position + target[1]], coords={
-                    "time": time_index.values, "dim_0": list(range(target[1]))}, dims=["time", "dim_0"])
+                    time_index.name: time_index.values,
+                    "dim_0": list(range(target[1]))}, dims=[time_index.name, "dim_0"])
                 position += target[1]
         return result